+++ src/qcodeedit/lib/document/qdocumentline_p.h
@@ -75,7 +75,7 @@ public:
 		int wrappedLineForCursor(int cpos) const;
 		int wrappedLineForCursorNoLock(int cpos) const;
 
-        int documentOffsetToCursor(int x, int y, bool disallowPositionBeyondLine = false) const;
+        int documentOffsetToCursor(qreal x, qreal y, bool disallowPositionBeyondLine = false) const;
 		void cursorToDocumentOffset(int cpos, qreal &x, qreal &y) const;
 
 		QPointF cursorToDocumentOffset(int cpos) const;
